+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/parser/clause/SelectListClauseParser.java
@@ -38,6 +38,9 @@ import io.shardingsphere.core.rule.ShardingRule;
 import io.shardingsphere.core.util.SQLUtil;
 import lombok.Getter;
 
+import java.util.Collection;
+import java.util.Collections;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -71,34 +74,51 @@ public abstract class SelectListClauseParser implements SQLClauseParser {
      */
     public void parse(final SelectStatement selectStatement, final List<SelectItem> items) {
         do {
-            selectStatement.getItems().add(parseSelectItem(selectStatement));
+            selectStatement.getItems().addAll(parseSelectItems(selectStatement));
         } while (lexerEngine.skipIfEqual(Symbol.COMMA));
         selectStatement.setSelectListLastPosition(lexerEngine.getCurrentToken().getEndPosition() - lexerEngine.getCurrentToken().getLiterals().length());
         items.addAll(selectStatement.getItems());
     }
     
-    private SelectItem parseSelectItem(final SelectStatement selectStatement) {
+    private Collection<SelectItem> parseSelectItems(final SelectStatement selectStatement) {
         lexerEngine.skipIfEqual(getSkippedKeywordsBeforeSelectItem());
-        return getSelectItem(selectStatement);
+        Collection<SelectItem> result = getSelectItems(selectStatement);
+        reviseDistinctSelectItems(selectStatement, result);
+        return result;
     }
     
-    private SelectItem getSelectItem(final SelectStatement selectStatement) {
-        final SelectItem result;
+    private Collection<SelectItem> getSelectItems(final SelectStatement selectStatement) {
+        final Collection<SelectItem> result = new LinkedList<>();
         if (isRowNumberSelectItem()) {
-            result = parseRowNumberSelectItem(selectStatement);
+            result.add(parseRowNumberSelectItem(selectStatement));
         } else if (isDistinctSelectItem()) {
-            result = parseDistinctSelectItem(selectStatement);
+            result.add(parseDistinctSelectItem(selectStatement));
+            addStarSelectItem(result);
             parseRestSelectItem(selectStatement);
         } else if (isStarSelectItem()) {
             selectStatement.setContainStar(true);
-            result = parseStarSelectItem();
+            result.add(parseStarSelectItem());
         } else if (isAggregationSelectItem()) {
-            result = parseAggregationSelectItem(selectStatement);
+            result.add(parseAggregationSelectItem(selectStatement));
             parseRestSelectItem(selectStatement);
         } else {
-            result = parseCommonOrStarSelectItem(selectStatement);
+            result.add(parseCommonOrStarSelectItem(selectStatement));
         }
         return result;
+    }
+    
+    private void addStarSelectItem(final Collection<SelectItem> result) {
+        if (isStarSelectItem()) {
+            result.add(parseStarSelectItem());
+        }
+    }
+    
+    private void reviseDistinctSelectItems(final SelectStatement selectStatement, final Collection<SelectItem> selectItems) {
+        for (SelectItem each : selectItems) {
+            if (!(selectStatement.getDistinctSelectItems().isEmpty() || each instanceof StarSelectItem)) {
+                selectStatement.getDistinctSelectItems().get(0).getDistinctColumnNames().add(each.getExpression());
+            }
+        }
     }
     
     protected abstract Keyword[] getSkippedKeywordsBeforeSelectItem();
@@ -114,9 +134,12 @@ public abstract class SelectListClauseParser implements SQLClauseParser {
     private SelectItem parseDistinctSelectItem(final SelectStatement selectStatement) {
         lexerEngine.nextToken();
         String distinctColumnName = lexerEngine.getCurrentToken().getLiterals();
+        if (Symbol.STAR == lexerEngine.getCurrentToken().getType()) {
+            return new DistinctSelectItem(Collections.<String>emptyList(), aliasExpressionParser.parseSelectItemAlias());
+        }
         lexerEngine.nextToken();
         distinctColumnName = SQLUtil.getExactlyValue(distinctColumnName + parseRestSelectItem(selectStatement));
-        return new DistinctSelectItem(distinctColumnName, aliasExpressionParser.parseSelectItemAlias());
+        return new DistinctSelectItem(Collections.singletonList(distinctColumnName), aliasExpressionParser.parseSelectItemAlias());
     }
     
     private boolean isStarSelectItem() {