+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/parser/sql/dql/select/AbstractSelectParser.java
@@ -222,7 +222,7 @@ public abstract class AbstractSelectParser implements SQLParser {
             return false;
         }
         DistinctSelectItem distinctSelectItem = (DistinctSelectItem) selectItem;
-        return orderItem.getColumnLabel().equals(distinctSelectItem.getDistinctColumnName()) || orderItem.getColumnLabel().equals(distinctSelectItem.getColumnLabel());
+        return distinctSelectItem.getDistinctColumnNames().contains(orderItem.getColumnLabel()) || orderItem.getColumnLabel().equals(distinctSelectItem.getColumnLabel());
     }
     
     private boolean isSameAlias(final SelectItem selectItem, final OrderItem orderItem) {